#ETR-1294 cambiando el campo nro_deposito a bigint

+++ modelo/MODProcesoCaja.php
@@ -208,7 +208,7 @@ class MODProcesoCaja extends MODbase{
 		$this->captura('desc_cuenta_bancaria','text');
 		$this->captura('fecha','date');
 		$this->captura('tipo','varchar');
-		$this->captura('nro_deposito','int4');
+		$this->captura('nro_deposito','bigint');
 		$this->captura('importe_deposito','numeric');
 		$this->captura('origen','varchar');
 		$this->captura('nombre_finalidad','varchar');
@@ -237,7 +237,7 @@ class MODProcesoCaja extends MODbase{
 		$this->setParametro('fecha','fecha','date');
 		$this->setParametro('tipo','tipo','varchar');
 		$this->setParametro('observaciones','observaciones','varchar');
-		$this->setParametro('nro_deposito','nro_deposito','numeric');
+		$this->setParametro('nro_deposito','nro_deposito','bigint');
 		$this->setParametro('importe_deposito','importe_deposito','numeric');
 		$this->setParametro('origen','origen','varchar');
 		$this->setParametro('tabla','tabla','varchar');
